+++ .vscode/settings.json
@@ -28,6 +28,7 @@
 		"HAEWON",
 		"HANNI",
 		"HYEIN",
+		"ILLIT",
 		"INFJ",
 		"INFP",
 		"ISFJ",
@@ -37,20 +38,20 @@
 		"JANG",
 		"Jingyu",
 		"JIWOO",
+		"josa",
 		"KAZUHA",
 		"LEESEO",
 		"mbti",
 		"MINJI",
 		"NINGNING",
 		"NMIXX",
-		"ILLIT",
-		"josa",
 		"Pretendard",
 		"SAKURA",
 		"slicktest",
 		"SSERAFIM",
 		"SULLYOON",
 		"Testpage",
+		"uuidv",
 		"WONYOUNG",
 		"YUJIN",
 		"YUNJIN"
