+++ src/pages/List/Chart/components/ChartVoteModal/VoteButton.jsx
@@ -1,9 +1,9 @@
 /** @jsxImportSource @emotion/react */
-import Button from "../../../../../components/Button/Button";
-import useIsMobile from "../../../../../hooks/useIsMobile";
+import Button from "@/components/Button/Button";
+import useIsMobile from "@/hooks/useIsMobile";
 import { CreditInfo, centerAlignStyle } from "./VoteButton.styles";
 
-export default function VoteButton({ onSubmit }) {
+export default function VoteButton({ isSubmitting }) {
 	const isMobile = useIsMobile();
 	const buttonSize = isMobile ? "vote-md" : "vote-lg";
 
@@ -13,9 +13,9 @@ export default function VoteButton({ onSubmit }) {
 				type="submit"
 				size={buttonSize}
 				fullWidth={true}
-				onClick={onSubmit}
+				disabled={isSubmitting}
 			>
-				투표하기
+				{isSubmitting ? "제출 중..." : "투표하기"}
 			</Button>
 			<p css={CreditInfo}>
 				투표하는 데 <span>1000 크레딧</span>이 소모됩니다.
