+++ src/pages/List/Chart/components/ChartVoteModal/index.jsx
@@ -1,7 +1,8 @@
+import LoadingError from "@/components/Error";
+import useSafeSubmit from "@/hooks/useSafeSubmit";
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
-import LoadingError from "../../../../../components/Error";
-import { useChartVoteModal } from "../../../../../hooks/useChartVoteModal";
+import { useChartVoteModal } from "../hooks/useChartVoteModal";
 import VoteIdolSkeleton from "./Skeleton";
 import VoteButton from "./VoteButton";
 import VoteIdolList from "./VoteIdolList";
@@ -15,8 +16,10 @@ export default function ChartVoteModal({
 	const { selectedIdolId, handleVote, handleIdolSelect, loading, error } =
 		useChartVoteModal(gender, closeModal, idols, setIdols);
 
+	const { isSubmitting, safeSubmit } = useSafeSubmit(handleVote);
+
 	return (
-		<form onSubmit={handleVote} css={VoteFormStyles}>
+		<form onSubmit={safeSubmit} css={VoteFormStyles}>
 			{loading ? (
 				<div>
 					<VoteIdolSkeleton />
@@ -30,7 +33,7 @@ export default function ChartVoteModal({
 					onSelectIdol={handleIdolSelect}
 				/>
 			)}
-			<VoteButton />
+			<VoteButton isSubmitting={isSubmitting} />
 		</form>
 	);
 }
